consolidates common functions of the tests into helper
