+++ network/topology/randomizedTopology_test.go
@@ -1,11 +1,9 @@
 package topology
 
 import (
-	"os"
 	"sort"
 	"testing"
 
-	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
@@ -13,7 +11,7 @@ import (
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/model/flow/filter"
 	"github.com/onflow/flow-go/network"
-	protocol2 "github.com/onflow/flow-go/state/protocol"
+	"github.com/onflow/flow-go/state/protocol"
 	"github.com/onflow/flow-go/utils/unittest"
 )
 
@@ -21,11 +19,11 @@ import (
 // topology that is needed for our network.
 type RandomizedTopologyTestSuite struct {
 	suite.Suite
-	state    protocol2.State   // represents a mocked protocol state
+	state    protocol.State    // represents a mocked protocol state
 	all      flow.IdentityList // represents the identity list of all nodes in the system
 	clusters flow.ClusterList  // represents list of cluster ids of collection nodes
 	subMngr  []network.SubscriptionManager
-	logger   zerolog.Logger
+	edgeProb float64
 }
 
 // TestRandomizedTopologyTestSuite starts all the tests in this test suite
@@ -40,6 +38,9 @@ func (suite *RandomizedTopologyTestSuite) SetupTest() {
 	nCollectors := 100
 	nTotal := 1000
 
+	// sets edge probability to 0.3
+	suite.edgeProb = 0.3
+
 	collectors := unittest.IdentityListFixture(nCollectors, unittest.WithRole(flow.RoleCollection))
 	others := unittest.IdentityListFixture(nTotal, unittest.WithAllRolesExcept(flow.RoleCollection))
 	suite.all = append(others, collectors...)
@@ -49,8 +50,6 @@ func (suite *RandomizedTopologyTestSuite) SetupTest() {
 		suite.all.Filter(filter.HasRole(flow.RoleCollection)), uint(nClusters))
 
 	suite.subMngr = MockSubscriptionManager(suite.T(), suite.all)
-
-	suite.logger = zerolog.New(os.Stderr).Level(zerolog.DebugLevel)
 }
 
 // TestUnhappyInitialization concerns initializing randomized topology with unhappy inputs
@@ -72,81 +71,15 @@ func (suite *RandomizedTopologyTestSuite) TestUnhappyInitialization() {
 	require.Error(suite.T(), err)
 }
 
-// TestTopologySize_Topic verifies that size of each topology fanout per topic is greater than
-// `(k+1)/2` where `k` is number of nodes subscribed to a topic. It does that over 100 random iterations.
-func (suite *RandomizedTopologyTestSuite) TestTopologySize_Topic() {
-	for i := 0; i < 100; i++ {
-		top, err := NewTopicBasedTopology(suite.all[0].NodeID, suite.logger, suite.state, suite.subMngr[0])
-		require.NoError(suite.T(), err)
-
-		topics := engine.ChannelIDsByRole(suite.all[0].Role)
-		require.Greater(suite.T(), len(topics), 1)
-
-		for _, topic := range topics {
-			// extracts total number of nodes subscribed to topic
-			roles, ok := engine.RolesByChannelID(topic)
-			require.True(suite.T(), ok)
-
-			ids, err := top.subsetChannel(suite.all, nil, topic)
-			require.NoError(suite.T(), err)
-
-			// counts total number of nodes that has the roles and are not `suite.me`  (node of interest).
-			total := len(suite.all.Filter(filter.And(filter.HasRole(roles...),
-				filter.Not(filter.HasNodeID(suite.all[0].NodeID)))))
-			require.True(suite.T(), float64(len(ids)) >= (float64)(total+1)/2)
-		}
-	}
-}
-
-// TestDeteministicity is a weak test that verifies the same seed generates the same topology for a topic.
-//
-// It also checks the topology against non-inclusion of the node itself in its own topology.
-func (suite *RandomizedTopologyTestSuite) TestDeteministicity() {
-	// creates a topology using the graph sampler
-	top, err := NewTopicBasedTopology(suite.all[0].NodeID, suite.logger, suite.state, suite.subMngr[0])
-	require.NoError(suite.T(), err)
-
-	topics := engine.ChannelIDsByRole(suite.all[0].Role)
-	require.Greater(suite.T(), len(topics), 1)
-
-	// for each topic samples 100 topologies
-	// all topologies for a topic should be the same
-	for _, topic := range topics {
-		var previous, current []string
-		for i := 0; i < 100; i++ {
-			previous = current
-			current = nil
-
-			// generate a new topology with a the same all, size and seed
-			ids, err := top.subsetChannel(suite.all, nil, topic)
-			require.NoError(suite.T(), err)
-
-			// topology should not contain the node itself
-			require.Empty(suite.T(), ids.Filter(filter.HasNodeID(suite.all[0].NodeID)))
-
-			for _, v := range ids {
-				current = append(current, v.NodeID.String())
-			}
-			// no guarantees about order is made by Topology.subsetChannel(), hence sort the return values before comparision
-			sort.Strings(current)
-
-			if previous == nil {
-				continue
-			}
-
-			// assert that a different seed generates a different topology
-			require.Equal(suite.T(), previous, current)
-		}
-	}
-}
-
-// TestUniqueness generates a topology for the first topic of consensus nodes.
+// TestUniqueness is a weak uniqueness check evaluates that no two consecutive connected components
+// on the same channel are the same.
+// It generates a topology for the first topic of consensus nodes.
 // Since topologies are seeded with the node ids, it evaluates that every two consecutive
 // topologies of the same topic for distinct nodes are distinct.
 //
 // It also checks the topology against non-inclusion of the node itself in its own topology.
 //
-// Note: currently we are using a linear fanout for guaranteed delivery, hence there are
+// Note: currently we are using a uniform probability, hence there are
 // C(n, (n+1)/2) many unique topologies for the same topic across different nodes. Even for small numbers
 // like n = 300, the potential outcomes are large enough (i.e., 10e88) so that the uniqueness is guaranteed.
 // This test however, performs a very weak uniqueness test by checking the uniqueness among consecutive topologies.
@@ -168,9 +101,9 @@ func (suite *RandomizedTopologyTestSuite) TestUniqueness() {
 		current = nil
 
 		// creates and samples a new topic aware topology for the first topic of consensus nodes
-		top, err := NewTopicBasedTopology(identity.NodeID, suite.logger, suite.state, suite.subMngr[i])
+		top, err := NewRandomizedTopology(identity.NodeID, suite.edgeProb, suite.state, suite.subMngr[i])
 		require.NoError(suite.T(), err)
-		ids, err := top.subsetChannel(suite.all, nil, topics[0])
+		ids, err := top.subsetChannel(suite.all, topics[0])
 		require.NoError(suite.T(), err)
 
 		// topology should not contain the node itself
@@ -190,7 +123,7 @@ func (suite *RandomizedTopologyTestSuite) TestUniqueness() {
 	}
 }
 
-// TestConnectedness_NonClusterTopics checks whether graph components corresponding to a
+// TestConnectedness_NonClusterChannelID checks whether graph components corresponding to a
 // non-cluster channel ID are individually connected.
 func (suite *RandomizedTopologyTestSuite) TestConnectedness_NonClusterChannelID() {
 	channelID := engine.TestNetwork
@@ -199,11 +132,11 @@ func (suite *RandomizedTopologyTestSuite) TestConnectedness_NonClusterChannelID(
 
 	for i, id := range suite.all {
 		// creates a topic-based topology for node
-		top, err := NewTopicBasedTopology(id.NodeID, suite.logger, suite.state, suite.subMngr[i])
+		top, err := NewRandomizedTopology(id.NodeID, suite.edgeProb, suite.state, suite.subMngr[i])
 		require.NoError(suite.T(), err)
 
 		// samples subset of topology
-		subset, err := top.subsetChannel(suite.all, nil, channelID)
+		subset, err := top.subsetChannel(suite.all, channelID)
 		require.NoError(suite.T(), err)
 
 		channelIDAdjMap[id.NodeID] = subset
@@ -223,191 +156,62 @@ func (suite *RandomizedTopologyTestSuite) TestConnectedness_ClusterChannelID() {
 
 	// iterates over collection nodes
 	for i, id := range suite.all.Filter(filter.HasRole(flow.RoleCollection)) {
-		// creates a channelID-based topology for node
-		top, err := NewTopicBasedTopology(id.NodeID, suite.logger, suite.state, suite.subMngr[i])
+		// creates a randomized topology for node
+		top, err := NewRandomizedTopology(id.NodeID, suite.edgeProb, suite.state, suite.subMngr[i])
 		require.NoError(suite.T(), err)
 
 		// samples subset of topology
-		subset, err := top.subsetChannel(suite.all, nil, channelID)
+		subset, err := top.subsetChannel(suite.all, channelID)
 		require.NoError(suite.T(), err)
+
+		uniquenessCheck(suite.T(), subset)
 
 		channelIDAdjMap[id.NodeID] = subset
 	}
 
 	// check that each of the collection clusters forms a connected graph
 	for _, cluster := range suite.clusters {
-		suite.checkConnectednessByCluster(suite.T(), channelIDAdjMap, cluster)
+		checkConnectednessByCluster(suite.T(), channelIDAdjMap, suite.all, cluster)
 	}
-}
-
-// TestLinearFanout_UnconditionalSampling evaluates that sampling a connected graph fanout
-// with an empty `shouldHave` list follows the LinearFanoutFunc,
-// and it also does not contain duplicate element.
-func (suite *RandomizedTopologyTestSuite) TestLinearFanout_UnconditionalSampling() {
-	// samples with no `shouldHave` set.
-	top, err := NewTopicBasedTopology(suite.all[0].NodeID, suite.logger, suite.state, suite.subMngr[0])
-	require.NoError(suite.T(), err)
-
-	sample, err := top.sampleConnectedGraph(suite.all, nil)
-	require.NoError(suite.T(), err)
-
-	// the LinearFanoutGraphSampler utilizes the LinearFanoutFunc. Hence any sample it makes should have
-	// the size equal to applying LinearFanoutFunc over the original set.
-	expectedFanout := LinearFanout(len(suite.all))
-	require.Equal(suite.T(), len(sample), expectedFanout)
-
-	// checks sample does not include any duplicate
-	suite.uniquenessCheck(sample)
-}
-
-// TestLinearFanout_ConditionalSampling evaluates that sampling a connected graph fanout with a shouldHave set
-// follows the LinearFanoutFunc, and it also does not contain duplicate element.
-func (suite *RandomizedTopologyTestSuite) TestLinearFanout_ConditionalSampling() {
-	// samples 10 all into `shouldHave` set.
-	shouldHave := suite.all.Sample(10)
-
-	// creates a topology for the node
-	top, err := NewTopicBasedTopology(suite.all[0].NodeID, suite.logger, suite.state, suite.subMngr[0])
-	require.NoError(suite.T(), err)
-
-	// samples a connected graph of `all` that includes `shouldHave` set.
-	sample, err := top.sampleConnectedGraph(suite.all, shouldHave)
-	require.NoError(suite.T(), err)
-
-	// the LinearFanoutGraphSampler utilizes the LinearFanoutFunc. Hence any sample it makes should have
-	// the size equal to applying LinearFanoutFunc over the original set.
-	expectedFanout := LinearFanout(len(suite.all))
-	require.Equal(suite.T(), len(sample), expectedFanout)
-
-	// checks sample does not include any duplicate
-	suite.uniquenessCheck(sample)
-
-	// checks inclusion of all shouldHave ones into sample
-	for _, id := range shouldHave {
-		require.Contains(suite.T(), sample, id)
-	}
-}
-
-// TestLinearFanout_SmallerAll evaluates that sampling a connected graph fanout with a shouldHave set
-// that is greater than required fanout, returns the `shouldHave` set instead.
-func (suite *RandomizedTopologyTestSuite) TestLinearFanout_SmallerAll() {
-	// samples 10 all into 'shouldHave'.
-	shouldHave := suite.all.Sample(10)
-	// samples a smaller component of all with 5 nodes and combines with `shouldHave`
-	smallerAll := suite.all.Filter(filter.Not(filter.In(shouldHave))).Sample(5).Union(shouldHave)
-
-	// creates a topology for the node
-	top, err := NewTopicBasedTopology(suite.all[0].NodeID, suite.logger, suite.state, suite.subMngr[0])
-	require.NoError(suite.T(), err)
-
-	// total size of smallerAll is 15, and it requires a linear fanout of 8 which is less than
-	// size of `shouldHave` set, so the shouldHave itself should return
-	sample, err := top.sampleConnectedGraph(smallerAll, shouldHave)
-	require.NoError(suite.T(), err)
-	require.Equal(suite.T(), len(sample), len(shouldHave))
-	require.ElementsMatch(suite.T(), sample, shouldHave)
-}
-
-// TestLinearFanout_SubsetViolence evaluates that trying to sample a connected graph when `shouldHave`
-// is not a subset of `all` returns an error.
-func (suite *RandomizedTopologyTestSuite) TestLinearFanout_SubsetViolence() {
-	// samples 10 all into 'shouldHave',
-	shouldHave := suite.all.Sample(10)
-	// excludes one of the `shouldHave` all from all, hence it is no longer a subset
-	excludedAll := suite.all.Filter(filter.Not(filter.HasNodeID(shouldHave[0].NodeID)))
-
-	// creates a topology for the node
-	top, err := NewTopicBasedTopology(suite.all[0].NodeID, suite.logger, suite.state, suite.subMngr[0])
-	require.NoError(suite.T(), err)
-
-	// since `shouldHave` is not a subset of `excludedAll` it should return an error
-	_, err = top.sampleConnectedGraph(excludedAll, shouldHave)
-	require.Error(suite.T(), err)
 }
 
 // TestLinearFanout_EmptyAllSet evaluates that trying to sample a connected graph when `all`
 // is empty returns an error.
 func (suite *RandomizedTopologyTestSuite) TestLinearFanout_EmptyAllSet() {
-	// samples 10 all into 'shouldHave'.
-	shouldHave := suite.all.Sample(10)
-
 	// creates a topology for the node
-	top, err := NewTopicBasedTopology(suite.all[0].NodeID, suite.logger, suite.state, suite.subMngr[0])
+	top, err := NewRandomizedTopology(suite.all[0].NodeID, suite.edgeProb, suite.state, suite.subMngr[0])
 	require.NoError(suite.T(), err)
 
-	// sampling with empty `all` and non-empty `shouldHave`
-	_, err = top.sampleConnectedGraph(flow.IdentityList{}, shouldHave)
+	// sampling with empty `all`
+	_, err = top.sampleFanout(flow.IdentityList{})
 	require.Error(suite.T(), err)
 
-	// sampling with empty all and nil `shouldHave`
-	_, err = top.sampleConnectedGraph(flow.IdentityList{}, nil)
+	// sampling with empty all
+	_, err = top.sampleFanout(flow.IdentityList{})
 	require.Error(suite.T(), err)
 
-	// sampling with nil all and nil `shouldHave`
-	_, err = top.sampleConnectedGraph(nil, nil)
+	// sampling with nil all
+	_, err = top.sampleFanout(nil)
 	require.Error(suite.T(), err)
 }
 
-// TestConnectedness_Unconditionally evaluates that samples returned by the LinearFanoutGraphSampler with
-// empty `shouldHave` constitute a connected graph.
-func (suite *RandomizedTopologyTestSuite) TestConnectedness_Unconditionally() {
+// TestSampleFanoutConnectedness evaluates that sampleFanout with default edge probability of this suite provides
+// a connected graph,
+func (suite *RandomizedTopologyTestSuite) TestSampleFanoutConnectedness() {
 	adjMap := make(map[flow.Identifier]flow.IdentityList)
 	for i, id := range suite.all {
 		// creates a topology for the node
-		top, err := NewTopicBasedTopology(id.NodeID, suite.logger, suite.state, suite.subMngr[i])
+		top, err := NewRandomizedTopology(id.NodeID, suite.edgeProb, suite.state, suite.subMngr[i])
 		require.NoError(suite.T(), err)
 
 		// samples a graph and stores it in adjacency map
-		sample, err := top.sampleConnectedGraph(suite.all, nil)
+		sample, err := top.sampleFanout(suite.all)
+		require.NoError(suite.T(), err)
+
+		uniquenessCheck(suite.T(), sample)
+
 		adjMap[id.NodeID] = sample
 	}
 
 	CheckGraphConnected(suite.T(), adjMap, suite.all, filter.In(suite.all))
-}
-
-// TestConnectedness_Conditionally evaluates that samples returned by the LinearFanoutGraphSampler with
-// some `shouldHave` constitute a connected graph.
-func (suite *RandomizedTopologyTestSuite) TestConnectedness_Conditionally() {
-	adjMap := make(map[flow.Identifier]flow.IdentityList)
-	for i, id := range suite.all {
-		// creates a topology for the node
-		top, err := NewTopicBasedTopology(id.NodeID, suite.logger, suite.state, suite.subMngr[i])
-		require.NoError(suite.T(), err)
-
-		// samples a graph and stores it in adjacency map
-		// sampling is done with a non-empty `shouldHave` subset of 10 randomly chosen all
-		shouldHave := suite.all.Sample(10)
-		sample, err := top.sampleConnectedGraph(suite.all, shouldHave)
-
-		// evaluates inclusion of should haves in sample
-		for _, shouldHaveID := range shouldHave {
-			require.Contains(suite.T(), sample, shouldHaveID)
-		}
-		adjMap[id.NodeID] = sample
-	}
-
-	CheckGraphConnected(suite.T(), adjMap, suite.all, filter.In(suite.all))
-}
-
-// uniquenessCheck is a test helper method that fails the test if all include any duplicate identity.
-func (suite *RandomizedTopologyTestSuite) uniquenessCheck(ids flow.IdentityList) {
-	seen := make(map[flow.Identity]struct{})
-	for _, id := range ids {
-		// checks if id is duplicate in ids list
-		_, ok := seen[*id]
-		require.False(suite.T(), ok)
-
-		// marks id as seen
-		seen[*id] = struct{}{}
-	}
-}
-
-// checkConnectednessByCluster is a test helper that checks all nodes belong to a cluster are connected.
-func (suite *RandomizedTopologyTestSuite) checkConnectednessByCluster(t *testing.T,
-	adjMap map[flow.Identifier]flow.IdentityList,
-	cluster flow.IdentityList) {
-	CheckGraphConnected(t,
-		adjMap,
-		suite.all,
-		filter.In(cluster))
 }
